Added --dry-run command and corrected issue numbering

+++ comicrock/main.py
@@ -36,7 +36,7 @@ class ComicRock(object):
             soup = self.get_html(url)
         return soup.select('td strong')[0].text
 
-    def download_series(self, url, start=1, end=-1):
+    def download_series(self, url, start=0, end=-1, dry_run=False):
         soup = self.get_html(url)
         book_name = self.get_book_name(url, soup=soup)
         chapter_list = [x['href'] for x in soup.select(self.chapter_selector)]
@@ -45,38 +45,38 @@ class ComicRock(object):
         serialized_book_name = url.rpartition('/')[-1]
         if end < 0:
             end = len(chapter_list)
-        print('This book has {} issues. Downloading from {} to {}'.format(len(chapter_list), start, end))
-        for no, chapter_url in enumerate(chapter_list):
-            chapter_no = no + 1
-            if no == 0 and 'chapter-0' in chapter_url:
-                chapter_no = no
-                start = 0
+        print('This book has {} issues.'.format(len(chapter_list)))
+        for chapter_url in chapter_list:
+            chapter_no = int(chapter_url.rpartition('/')[-1].split('-')[-1])
             if start <= chapter_no <= end:
-                self.download_issue(chapter_no, chapter_url, book_name, serialized_book_name, book_path)
+                self.download_issue(chapter_no, chapter_url, book_name, serialized_book_name, book_path, dry_run)
         return book_path
 
-    def download_issue(self, no, url, book_name, serialized_book_name, book_path):
+    def download_issue(self, no, url, book_name, serialized_book_name, book_path, dry_run):
         soup = self.get_html(url)
         page_count = soup.select('div.ct-right div.label')[0].text
         page_count = [int(s) for s in page_count.split() if s.isdigit()][0]
         issue_name = '{name}-{no}'.format(name=book_name, no=no)
         issue_path = os.path.join(book_path, issue_name)
         archive_path = os.path.join(book_path, issue_name)
-        if os.path.exists(archive_path+'.cbz'):
-            return 0
-        os.makedirs(issue_path, exist_ok=True)
-        for page in range(1, page_count+1):
-            print('Downloading Issue #{} Page {} of {}'.format(no, page, page_count), end='\r')
-            page_url = urljoin(self.image_url, '{book}/{chapter}/{page}.jpg'.format(book=serialized_book_name,
-                                                                                    chapter=no, page=page))
-            page_path = os.path.join(issue_path, '{0:03d}.jpg'.format(page))
-            try:
-                self.download_image(page_url, page_path)
-            except Exception:
-                print('Error on saving page {} of {} issue {}'.format(page, book_name, no))
-        shutil.make_archive(archive_path, 'zip', issue_path)
-        os.rename(archive_path+'.zip', archive_path+'.cbz')
-        shutil.rmtree(issue_path)
+        if not dry_run:
+            if os.path.exists(archive_path+'.cbz'):
+                return 0
+            os.makedirs(issue_path, exist_ok=True)
+            for page in range(1, page_count+1):
+                print('Downloading Issue #{} Page {} of {}'.format(no, page, page_count), end='\r')
+                page_url = urljoin(self.image_url, '{book}/{chapter}/{page}.jpg'.format(book=serialized_book_name,
+                                                                                        chapter=no, page=page))
+                page_path = os.path.join(issue_path, '{0:03d}.jpg'.format(page))
+                try:
+                    self.download_image(page_url, page_path)
+                except Exception:
+                    print('Error on saving page {} of {} issue {}'.format(page, book_name, no))
+            shutil.make_archive(archive_path, 'zip', issue_path)
+            os.rename(archive_path+'.zip', archive_path+'.cbz')
+            shutil.rmtree(issue_path)
+        else:
+            print('Issue #{}, URL: {}, Pages: {}'.format(no, url, page_count))
 
     def download_image(self, url, path):
         r = requests.get(url)
